requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
Django==5.2.8
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
sqlparse==0.5.4
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
